Add models.py and cancel button.
Slightly modularizes by using a separate models file. Also adds a cancel button for any post editing pages.
