+++ main.py
@@ -24,7 +24,7 @@ import hashlib
 import hmac
 import logging
 
-from google.appengine.ext import db
+import models
 
 template_dir = os.path.join(os.path.dirname(__file__), 'templates')
 jinja_env = jinja2.Environment(loader = jinja2.FileSystemLoader(template_dir),
@@ -112,362 +112,6 @@ def validate_password(username, password, hsh):
     salt = hsh.split('|')[1]
     return hsh == make_pw_hash(username, password, salt)
 
-
-# Keys
-def blog_key(name = 'default'):
-    return db.Key.from_path('blogs', name)
-
-def users_key(group = 'default'):
-    return db.Key.from_path('users', group)
-
-def comments_key(group = 'default'):
-    return db.Key.from_path('comments', group)
-
-def likes_key(group = 'default'):
-    return db.Key.from_path('likes', group)
-
-def dislikes_key(group = 'default'):
-    return db.Key.from_path('dislikes', group)
-
-# Database models
-class BlogPosts(db.Model):
-    """
-    Database model for BlogPosts. Also contains many related functions
-    that interact with other databases. Does the bulk of heavy lifting
-    in terms of rendering blog posts, likes, and dislikes.
-    """
-
-    subject = db.StringProperty(required = True)
-    content = db.TextProperty(required = True)
-    created = db.DateTimeProperty(auto_now_add = True)
-    modified = db.DateTimeProperty(auto_now = True)
-    created_user_id = db.StringProperty(required = True)
-
-    def render(self, user, is_static=False):
-        """
-        Renders the page. Passes through user and is_static for
-        other renderers to use.
-        """
-        return render_str("post.html",
-                          post = self,
-                          user = user,
-                          is_static=is_static)
-
-    @classmethod
-    def by_id(cls, post_id):
-        """Returns the BlogPost entity with the given post_id"""
-        return BlogPosts.get_by_id(int(post_id), parent = blog_key())
-
-    @classmethod
-    def get_post_creator_name(cls, user_id):
-        """
-        Returns the username of the given user_id that created the
-        given post_id.
-        """
-        user = Users.by_id(int(user_id))
-        return user.name
-
-    @classmethod
-    def get_post_comments(cls, post_id):
-        """Returns the comments with the given post id."""
-        return Comments.all().ancestor(comments_key()).filter(
-            'related_post_id =', str(post_id))
-
-    @classmethod
-    def get_top_comments(cls, post_id):
-        """Returns the top three comments for the given post_id."""
-        return Comments.all().ancestor(comments_key()).filter(
-            'related_post_id =', str(post_id)).run(limit=3)
-
-    @classmethod
-    def add_like_or_dislike(cls, like, dislike, post_id, user_id):
-        """
-        Takes the results of posting Like, Dislike, user_id, and
-        post_id. Adds a like or dislike depending which exists in
-        what's passed. Returns a dictionary composed of error and
-        post id.
-        """
-        # Get associated post
-        post = BlogPosts.by_id(int(post_id))
-        error = {}
-
-        if like:
-            if post.has_user_liked_post(user_id, post_id):
-                error["text"] = "You can't like a post twice."
-                error["post"] = long(post_id)
-            else:
-                like = Likes(parent = likes_key(), 
-                             post_id = post_id,
-                             user_id = user_id)
-                like.put()
-            if post.has_user_disliked_post(user_id, post_id):
-                has_disliked = post.get_dislike_by_user(user_id, post_id)
-                has_disliked.delete()
-        if dislike:
-            if post.has_user_disliked_post(user_id, post_id):
-                error["text"] = "You can't dislike a post twice."
-                error["post"] = long(post_id)
-            else:
-                dislike = Dislikes(parent = dislikes_key(),
-                                   post_id = post_id,
-                                   user_id = user_id)
-                dislike.put()
-            if post.has_user_liked_post(user_id, post_id):
-                has_liked = post.get_like_by_user(user_id, post_id)
-                has_liked.delete()
-
-        return error
-
-    @classmethod
-    def get_post_likes(cls, post_id):
-        """
-        Returns the filter for likes associated with the post. Use
-        `.run()` on the result.
-        """
-        return Likes.all().ancestor(likes_key()).filter(
-            'post_id =', str(post_id))
-
-    @classmethod
-    def get_post_dislikes(cls, post_id):
-        """
-        Returns the filter for dislikes associated with the post. Use
-        `.run()` on the result.
-        """
-        return Dislikes.all().ancestor(dislikes_key()).filter(
-            'post_id =', str(post_id))
-
-    @classmethod
-    def get_num_likes(cls, post_id):
-        """Returns the number of likes."""
-        likes = cls.get_post_likes(post_id)
-        count = 0
-        for like in likes:
-            if str(like.post_id) == str(post_id):
-                count += 1
-        return count
-
-    @classmethod
-    def get_num_dislikes(cls, post_id):
-        """Returns the number dislikes."""
-        dislikes = cls.get_post_dislikes(post_id)
-        count = 0
-        for dislike in dislikes:
-            if str(dislike.post_id) == str(post_id):
-                count += 1
-        return count
-
-    @classmethod
-    def has_user_liked_post(cls, user_id, post_id):
-        """
-        Returns True if the given user has created a like for the
-        given post and False otherwise.
-        """
-        likes_with_post_id = cls.get_post_likes(post_id)
-        has_liked = False
-        for like in likes_with_post_id.run():
-            if str(like.user_id) == str(user_id):
-                has_liked = True
-                break
-        return has_liked
-
-    @classmethod
-    def has_user_disliked_post(cls, user_id, post_id):
-        """
-        Returns True if the given user has created a dislike the given
-        post and False otherwise.
-        """
-        dislikes_with_post_id = cls.get_post_dislikes(post_id)
-        has_disliked = False
-        for dislike in dislikes_with_post_id.run():
-            if str(dislike.user_id) == str(user_id):
-                has_disliked = True
-                break
-        return has_disliked
-
-
-    @classmethod
-    def get_like_by_user(cls, user_id, post_id):
-        """
-        Returns a Likes entity with the given user_id for the
-        given post_id.
-        """
-        likes = cls.get_post_likes(post_id).run()
-        for like in likes:
-            if str(like.user_id) == str(user_id):
-                return like
-
-    @classmethod
-    def get_dislike_by_user(cls, user_id, post_id):
-        """
-        Returns a Dislikes entity with the given user_id for the
-        given post_id.
-        """
-        dislikes = cls.get_post_dislikes(post_id).run()
-        for dislike in dislikes:
-            if str(dislike.user_id) == str(user_id):
-                return dislike
-
-    @classmethod
-    def render_like(cls, post, disabled=False):
-        """
-        Renders the like for the post. If disabled is True, renders
-        the disabled version.
-        """
-        return render_str('like.html', disabled = disabled, post = post)
-
-    @classmethod
-    def render_dislike(cls, post, disabled=False):
-        """
-        Renders the dislike for the post. If disabled is True, renders
-        the disabled version.
-        """
-        return render_str('dislike.html', disabled = disabled, post = post)
-
-class Comments(db.Model):
-    """
-    Database model for Comments. Also contains several related
-    functions, some of which interact with other models.
-    """
-    subject = db.StringProperty(required = True)
-    content = db.TextProperty(required = True)
-    created_user_id = db.StringProperty(required = True)
-    related_post_id = db.StringProperty(required = True)
-    created = db.DateTimeProperty(auto_now_add = True)
-
-    @classmethod
-    def by_id(cls, comment_id):
-        """Returns the Comments entity with the given comment_id."""
-        return Comments.get_by_id(int(comment_id), parent = comments_key())
-
-    @classmethod
-    def add_comment(cls, subject, content, post_id, user_id):
-        """Adds a comment for the given post."""
-        if subject and content and post_id and user_id:
-            comment = Comments(parent = comments_key(),
-                               subject = subject,
-                               content = content,
-                               created_user_id = user_id,
-                               related_post_id = post_id)
-            comment.put()
-
-    @classmethod
-    def get_creator_username(cls, comment_id = ""):
-        """
-        Gets the creator username of the given comment id. If no
-        comment ID is given, attempts to use the current comment's
-        created_user_id.
-        """
-        if not comment_id:
-            comment_id = cls.key().id()
-        comment = cls.by_id(int(comment_id))
-        user = Users.by_id(int(comment.created_user_id))
-        return user.name
-
-    @classmethod
-    def render_comment(cls, comment_id, user, is_static=False):
-        """
-        Renders the given comment for the given post. If no comment ID
-        is given, attempts to render the current comment. If small is
-        True renders the condensed version for pages with multipl
-        posts.
-        """
-        if not is_static:
-            comment_renderer = "comment_sml.html"
-        else:
-            comment_renderer = "comment.html"
-        return render_str(
-            comment_renderer, comment = cls.by_id(int(comment_id)), user=user)
-
-class Likes(db.Model):
-    """
-    Database model for Likes. Simply contains likes and is called
-    by other classes.
-    """
-    post_id = db.StringProperty(required = True)
-    user_id = db.StringProperty(required = True)
-
-class Dislikes(db.Model):
-    """
-    Database model for Dislikes. Simply contains dislikes and is
-    called by other classes.
-    """
-    post_id = db.StringProperty(required = True)
-    user_id = db.StringProperty(required = True)
-
-class Users(db.Model):
-    """Users database model."""
-    
-    name = db.StringProperty(required = True)
-    password = db.StringProperty(required = True)
-    email = db.StringProperty()
-    created = db.DateTimeProperty(auto_now_add = True)
-    
-    @classmethod
-    def by_id(cls, uid):
-        """Returns the user with the given id."""
-        return Users.get_by_id(uid, parent = users_key())
-
-    @classmethod
-    def by_name(cls, name):
-        """
-        Returns the user with the given username. It's preferred to
-        use id when possible.
-        """
-        user = Users.all().filter('name =', name).get()
-        return user
-
-    @classmethod
-    def register(cls, name, password, email = None):
-        """
-        Calls registering the given data as a User entity. Still need
-        to put() after calling.
-        """
-        return Users(parent = users_key(),
-                     name = name,
-                     password = password,
-                     email = email)
-
-    @classmethod
-    def user_login(cls, name, password):
-        """Validates user's password given the username."""
-        logging.info('username = %s' % name)
-        logging.info('password = %s' % password)
-        user = cls.by_name(name)
-        logging.info('user = %s' % type(user))
-        if user and validate_password(name, password, user.password):
-            return user
-
-    @classmethod
-    def user_id_str(cls, user_id):
-        """
-        Returns a string version of the user_id. Used for comparing
-        strings in html.
-        """
-        return str(user_id)
-
-    @classmethod
-    def get_users_posts(cls, user_id):
-        """Returns the posts of the given user_id."""
-        users_posts = BlogPosts.all().ancestor(blog_key()).filter(
-            "created_user_id =", str(user_id))
-        if users_posts.get() != None:
-            posts = users_posts.run()
-        else:
-            posts = None
-        return posts
-
-    @classmethod
-    def render_post_editor(cls, subject = "", content = "", error = ""):
-        """
-        Renders the page editor. For new posts or comments and post
-        or comment editing.
-        """
-        return render_str('post_editor.html',
-                          subject = subject,
-                          content = content,
-                          error = error)
-
-
 # Page handlers
 class Handler(webapp2.RequestHandler):
     """Base Handler class. Has writing and rendering basics."""
@@ -513,7 +157,7 @@ class BlogHandler(Handler):
         previous_page = self.read_secure_cookie("page")
 
         if not previous_page:
-            previous_page = ""
+            previous_page = "/"
 
         return previous_page
 
@@ -561,6 +205,18 @@ class BlogHandler(Handler):
         else:
             pass
 
+    def check_cancel_edit(self):
+        """
+        Performs cancelling edit logic.
+
+        Checks if cancel was sent. If so, gets the previous page, then
+        redirects to it.
+        """
+        cancel = self.request.get("cancel")
+
+        if cancel:
+            self.redirect(str(self.get_previous_page()))
+
     def get_post_data(self):
         """Gets the data in the post needed for"""
 
@@ -582,14 +238,14 @@ class BlogHandler(Handler):
         """If user is logged in allows the handler to call the user."""
         webapp2.RequestHandler.initialize(self, *args, **kwargs)
         uid = self.read_secure_cookie('user_id')
-        self.user = uid and Users.by_id(int(uid))
+        self.user = uid and models.Users.by_id(int(uid))
 
 class BlogFrontPageHandler(BlogHandler):
     """Handler for './blog'. """
     def render_front(self, error=""):
         """Gets the 10 most recent posts and sends them blog_front."""
-        posts = BlogPosts.all().ancestor(blog_key()).order('-created').fetch(
-            limit=10)
+        posts = models.BlogPosts.all().ancestor(models.blog_key()).order(
+            '-created').fetch(limit=10)
 
         self.render("blog_front.html",
                     user=self.user,
@@ -611,10 +267,9 @@ class BlogFrontPageHandler(BlogHandler):
         error = {}
 
         if data["like"] or data["dislike"]:
-            error = BlogPosts.add_like_or_dislike(like=data["like"],
-                                                  dislike=data["dislike"],
-                                                  post_id=data["post_id"],
-                                                  user_id=data["user_id"])
+            error = models.BlogPosts.add_like_or_dislike(
+                like=data["like"], dislike=data["dislike"],
+                post_id=data["post_id"], user_id=data["user_id"])
             self.render_front(error)
         elif data["edit_post"] or data["delete_post"] or data["add_comment"]:
             self.edit_post_redirect(edit_post=data["edit_post"],
@@ -658,11 +313,14 @@ class NewPostPageHandler(BlogHandler):
         subject = self.request.get("subject")
         content = self.request.get("content")
 
+        self.check_cancel_edit()
+
         if subject and content:
-            post = BlogPosts(parent = blog_key(),
-                             subject = subject,
-                             content = content,
-                             created_user_id = str(self.user.key().id()))
+            post = models.BlogPosts(parent = models.blog_key(),
+                                    subject = subject,
+                                    content = content,
+                                    created_user_id = (
+                                        str(self.user.key().id())))
             post.put()
 
             self.redirect("/%s" % post.key().id())
@@ -678,7 +336,7 @@ class StaticPostPageHandler(BlogHandler):
         If the page does not renders the page with a simple 404
         message.
         """
-        post = BlogPosts.by_id(int(post_id))
+        post = models.BlogPosts.by_id(int(post_id))
         if post == None:
             error = {
                 "text" : "Sorry, we can't find that post.",
@@ -704,10 +362,9 @@ class StaticPostPageHandler(BlogHandler):
         error = {}
 
         if data["like"] or data["dislike"]:
-            error = BlogPosts.add_like_or_dislike(like=data["like"],
-                                                  dislike=data["dislike"],
-                                                  post_id=data["post_id"],
-                                                  user_id=data["user_id"])
+            error = models.BlogPosts.add_like_or_dislike(
+                like=data["like"], dislike=data["dislike"],
+                post_id=data["post_id"], user_id=data["user_id"])
             self.render_static_page(post_id=data["post_id"], error = error)
         elif data["edit_post"] or data["delete_post"] or data["add_comment"]:
             self.edit_post_redirect(edit_post=data["edit_post"],
@@ -741,8 +398,8 @@ class UserSignUpHandler(BlogHandler):
         page. Otherwise, render the page with errors.
         """
         # Create users db
-        users = db.GqlQuery("SELECT * FROM Users "
-                             "ORDER BY created DESC ")
+        users = models.db.GqlQuery("SELECT * FROM Users "
+                                   "ORDER BY created DESC ")
 
         # Get user info
         self.username = self.request.get('username')
@@ -779,15 +436,15 @@ class UserSignUpHandler(BlogHandler):
             self.render('signup.html', **params)
         else:
             # Check if the username entered already exists
-            user = Users.by_name(self.username)
+            user = models.Users.by_name(self.username)
             if user:
                 params['username_error'] = "That user already exists."
                 self.render('signup.html', **params)
             else:
                 # Create the user in the db
-                user = Users.register(self.username,
-                                      self.hashed_password,
-                                      self.email)
+                user = models.Users.register(self.username,
+                                             self.hashed_password,
+                                             self.email)
                 user.put()
 
                 self.page_login(user)
@@ -858,7 +515,7 @@ class DeletePostPage(BlogHandler):
         previous_page = self.get_previous_page()
 
         if post_id:
-            post = BlogPosts.by_id(int(post_id))
+            post = models.BlogPosts.by_id(int(post_id))
             if not str(self.user.key().id()) == str(post.created_user_id):
                 self.redirect('/')
             self.render('deletepost.html', user = self.user, post_id = post_id)
@@ -888,7 +545,7 @@ class DeletePostPage(BlogHandler):
             # If the user came from the post page, redirect to the front page
             if previous_page == post_id:
                 previous_page = "/"
-            post = BlogPosts.by_id(int(post_id))
+            post = models.BlogPosts.by_id(int(post_id))
             post.delete()
 
         # delete previous cookies
@@ -915,7 +572,7 @@ class EditPageHandler(BlogHandler):
 
         if not subject or not content:
             if post_id:
-                post = BlogPosts.by_id(int(post_id))
+                post = models.BlogPosts.by_id(int(post_id))
                 if not str(self.user.key().id()) == str(post.created_user_id):
                     self.redirect('/%s' % previous_page)
                 else:
@@ -952,9 +609,11 @@ class EditPageHandler(BlogHandler):
         post_id = self.read_secure_cookie("post_id")
         previous_page = self.get_previous_page()
 
+        self.check_cancel_edit()
+
         if subject and content:
             if post_id:
-                post = BlogPosts.by_id(int(post_id))
+                post = models.BlogPosts.by_id(int(post_id))
                 post.subject = subject
                 post.content = content
                 post.put()
@@ -976,7 +635,7 @@ class NewCommentPageHandler(BlogHandler):
         """Renders the new comment page with the given content."""
         post_id = self.read_secure_cookie("post_id")
         if post_id:
-            post = BlogPosts.by_id(int(post_id))
+            post = models.BlogPosts.by_id(int(post_id))
             if not post:
                 self.redirect('/')
             else:
@@ -1010,18 +669,20 @@ class NewCommentPageHandler(BlogHandler):
         post_id = self.read_secure_cookie("post_id")
         previous_page = self.get_previous_page()
 
+        self.check_cancel_edit()
+
         if subject and content:
-            Comments.add_comment(subject=subject,
-                                 content=content,
-                                 post_id=post_id,
-                                 user_id=str(self.user.key().id()))
+            models.Comments.add_comment(subject=subject,
+                                        content=content,
+                                        post_id=post_id,
+                                        user_id=str(self.user.key().id()))
 
             self.set_secure_cookie("post_id", "")
             self.set_secure_cookie("page", "")
             self.redirect("/%s" % previous_page)
         else:
             error = "We need both a subject and some content!"
-            self.render_new_post(subject, content, error)
+            self.render_new_comment_page(subject, content, error)
 
 class EditCommentHandler(BlogHandler):
     """Handler for the './editcomment' page."""
@@ -1043,7 +704,7 @@ class EditCommentHandler(BlogHandler):
         
         if not subject or not content:
             if comment_id:
-                comment = Comments.by_id(int(comment_id))
+                comment = models.Comments.by_id(int(comment_id))
                 if comment:
                     if not str(self.user.key().id()) == str(
                         comment.created_user_id):
@@ -1083,11 +744,13 @@ class EditCommentHandler(BlogHandler):
         comment_id = self.read_secure_cookie("comment_id")
         previous_page = self.get_previous_page()
 
+        self.check_cancel_edit()
+
         error = "We need both a subject and some content!"
 
         if subject and content:
             if comment_id:
-                comment = Comments.by_id(int(comment_id))
+                comment = models.Comments.by_id(int(comment_id))
                 comment.subject = subject
                 comment.content = content
                 comment.put()
@@ -1116,7 +779,7 @@ class DeleteCommentPageHandler(BlogHandler):
         previous_page = self.get_previous_page()
 
         if comment_id:
-            comment = Comments.by_id(int(comment_id))
+            comment = models.Comments.by_id(int(comment_id))
             if comment:
                 if not (
                     str(self.user.key().id()) == str(comment.created_user_id)):
@@ -1146,7 +809,7 @@ class DeleteCommentPageHandler(BlogHandler):
         previous_page = self.get_previous_page()
 
         if resp_yes:
-            comment = Comments.by_id(int(comment_id))
+            comment = models.Comments.by_id(int(comment_id))
             comment.delete()
 
         # delete previous cookies
@@ -1178,9 +841,6 @@ class LoginPageHandler(BlogHandler):
         """
         self.username = self.request.get('username')
         self.password = self.request.get('password')
-        logging.info("attempting login")
-        logging.info('username is %s' % self.username)
-        logging.info('password is %s' % self.password)
 
         user = Users.user_login(self.username, self.password)
